Enum mpl_metaplex::instruction::MetaplexInstruction

pub enum MetaplexInstruction {

    DeprecatedInitAuctionManagerV1(AuctionManagerSettingsV1),
    DeprecatedValidateSafetyDepositBoxV1,
    RedeemBid,
    RedeemFullRightsTransferBid,
    DeprecatedRedeemParticipationBid,
    StartAuction,
    ClaimBid,
    EmptyPaymentAccount(EmptyPaymentAccountArgs),
    SetStore(SetStoreArgs),
    SetWhitelistedCreator(SetWhitelistedCreatorArgs),
    DeprecatedValidateParticipation,
    DeprecatedPopulateParticipationPrintingAccount,
    RedeemUnusedWinningConfigItemsAsAuctioneer(RedeemUnusedWinningConfigItemsAsAuctioneerArgs),
    DecommissionAuctionManager,
    RedeemPrintingV2Bid(RedeemPrintingV2BidArgs),
    WithdrawMasterEdition,
    DeprecatedRedeemParticipationBidV2,
    InitAuctionManagerV2(InitAuctionManagerV2Args),
    ValidateSafetyDepositBoxV2(SafetyDepositConfig),
    RedeemParticipationBidV3(RedeemParticipationBidV3Args),
    EndAuction(EndAuctionArgs),
    SetStoreIndex(SetStoreIndexArgs),
    SetAuctionCache,
    SetStoreV2(SetStoreV2Args),

}

Instructions supported by the Fraction program.

Variants

DeprecatedInitAuctionManagerV1(AuctionManagerSettingsV1)

Initializes an Auction Manager V1

[writable] Uninitialized, unallocated auction manager account with pda of [‘metaplex’, auction_key from auction referenced below]

[] Combined vault account with authority set to auction manager account (this will be checked) Note in addition that this vault account should have authority set to this program’s pda of [‘metaplex’, auction_key]

[] Auction with auctioned item being set to the vault given and authority set to this program’s pda of [‘metaplex’, auction_key]

[] Authority for the Auction Manager

[signer] Payer

[] Accept payment account of same token mint as the auction for taking payment for open editions, owner should be auction manager key

[] Store that this auction manager will belong to

[] System sysvar

[] Rent sysvar

Tuple Fields of DeprecatedInitAuctionManagerV1

0: AuctionManagerSettingsV1

DeprecatedValidateSafetyDepositBoxV1

Validates that a given safety deposit box has in it contents that match the expected WinningConfig in the auction manager. A stateful call, this will error out if you call it a second time after validation has occurred. 0. [writable] Uninitialized Safety deposit validation ticket, pda of seed [‘metaplex’, program id, auction manager key, safety deposit key]

[writable] Auction manager

[writable] Metadata account

[writable] Original authority lookup - unallocated uninitialized pda account with seed [‘metaplex’, auction key, metadata key] We will store original authority here to return it later.

[] A whitelisted creator entry for the store of this auction manager pda of [‘metaplex’, store key, creator key] where creator key comes from creator list of metadata, any will do

[] The auction manager’s store key

[] Safety deposit box account

[] Safety deposit box storage account where the actual nft token is stored

[] Mint account of the token in the safety deposit box

[] Edition OR MasterEdition record key Remember this does not need to be an existing account (may not be depending on token), just is a pda with seed of [‘metadata’, program id, Printing mint id, ‘edition’]. - remember PDA is relative to token metadata program.

[] Vault account

[signer] Authority

[signer optional] Metadata Authority - Signer only required if doing a full ownership txfer

[signer] Payer

[] Token metadata program

[] System

[] Rent sysvar

[writable] Limited edition Printing mint account (optional - only if using sending Limited Edition)

[signer] Limited edition Printing mint Authority account, this will TEMPORARILY TRANSFER MINTING AUTHORITY to the auction manager until all limited editions have been redeemed for authority tokens.

RedeemBid

NOTE: Requires an AuctionManagerV1. Note: This requires that auction manager be in a Running state.

If an auction is complete, you can redeem your bid for a specific item here. If you are the first to do this, The auction manager will switch from Running state to Disbursing state. If you are the last, this may change the auction manager state to Finished provided that no authorities remain to be delegated for Master Edition tokens.

NOTE: Please note that it is totally possible to redeem a bid 2x - once for a prize you won and once at the RedeemParticipationBid point for an open edition that comes as a ‘token of appreciation’ for bidding. They are not mutually exclusive unless explicitly set to be that way.

[writable] Auction manager

[writable] Safety deposit token storage account

[writable] Destination account.

[writable] Bid redemption key - Just a PDA with seed [‘metaplex’, auction_key, bidder_metadata_key] that we will allocate to mark that you redeemed your bid

[writable] Safety deposit box account

[writable] Vault account

[writable] Fraction mint of the vault

[] Auction

[] Your BidderMetadata account

[signer optional] Your Bidder account - Only needs to be signer if payer does not own

[signer] Payer

[] Token program

[] Token Vault program

[] Token metadata program

[] Store

[] System

[] Rent sysvar

[] PDA-based Transfer authority to move the tokens from the store to the destination seed [‘vault’, program_id, vault key] but please note that this is a PDA relative to the Token Vault program, with the ‘vault’ prefix

[optional/writable] Master edition (if Printing type of WinningConfig)

[optional/writable] Reservation list PDA [‘metadata’, program id, master edition key, ‘reservation’, auction manager key] relative to token metadata program (if Printing type of WinningConfig)

[] Safety deposit config pda of [‘metaplex’, program id, auction manager, safety deposit] This account will only get used AND BE REQUIRED in the event this is an AuctionManagerV2

[] Auction extended (pda relative to auction of [‘auction’, program id, vault key, ‘extended’])

RedeemFullRightsTransferBid

Note: This requires that auction manager be in a Running state.

If an auction is complete, you can redeem your bid for the actual Master Edition itself if it’s for that prize here. If you are the first to do this, the auction manager will switch from Running state to Disbursing state. If you are the last, this may change the auction manager state to Finished provided that no authorities remain to be delegated for Master Edition tokens.

NOTE: Please note that it is totally possible to redeem a bid 2x - once for a prize you won and once at the RedeemParticipationBid point for an open edition that comes as a ‘token of appreciation’ for bidding. They are not mutually exclusive unless explicitly set to be that way.

[writable] Auction manager

[writable] Safety deposit token storage account

[writable] Destination account.

[writable] Bid redemption key - Just a PDA with seed [‘metaplex’, auction_key, bidder_metadata_key] that we will allocate to mark that you redeemed your bid

[writable] Safety deposit box account

[writable] Vault account

[writable] Fraction mint of the vault

[] Auction

[] Your BidderMetadata account

[signer optional] Your Bidder account - Only needs to be signer if payer does not own

[signer] Payer

[] Token program

[] Token Vault program

[] Token metadata program

[] Store

[] System

[] Rent sysvar

[writable] Master Metadata account (pda of [‘metadata’, program id, Printing mint id]) - remember PDA is relative to token metadata program (This account is optional, and will only be used if metadata is unique, otherwise this account key will be ignored no matter it’s value)

[] New authority for Master Metadata - If you are taking ownership of a Master Edition in and of itself, or a Limited Edition that isn’t newly minted for you during this auction ie someone else had it minted for themselves in a prior auction or through some other means, this is the account the metadata for these tokens will be delegated to after this transaction. Otherwise this account will be ignored.

[] PDA-based Transfer authority to move the tokens from the store to the destination seed [‘vault’, program_id, vault key] but please note that this is a PDA relative to the Token Vault program, with the ‘vault’ prefix

[] Safety deposit config pda of [‘metaplex’, program id, auction manager, safety deposit] This account will only get used AND BE REQUIRED in the event this is an AuctionManagerV2

[] Auction extended (pda relative to auction of [‘auction’, program id, vault key, ‘extended’])

DeprecatedRedeemParticipationBid

Note: This requires that auction manager be in a Running state.

If an auction is complete, you can redeem your bid for an Open Edition token if it is eligible. If you are the first to do this, The auction manager will switch from Running state to Disbursing state. If you are the last, this may change the auction manager state to Finished provided that no authorities remain to be delegated for Master Edition tokens.

NOTE: Please note that it is totally possible to redeem a bid 2x - once for a prize you won and once at this end point for a open edition that comes as a ‘token of appreciation’ for bidding. They are not mutually exclusive unless explicitly set to be that way.

NOTE: If you are redeeming a newly minted Open Edition, you must actually supply a destination account containing a token from a brand new mint. We do not provide the token to you. Our job with this action is to christen this mint + token combo as an official Open Edition.

[writable] Auction manager

[writable] Safety deposit token storage account

[writable] Destination account for limited edition authority token. Must be same mint as master edition Printing mint.

[writable] Bid redemption key - Just a PDA with seed [‘metaplex’, auction_key, bidder_metadata_key] that we will allocate to mark that you redeemed your bid

[] Safety deposit box account

[] Vault account

[] Safety deposit config pda of [‘metaplex’, program id, auction manager, safety deposit] This account will only get used in the event this is an AuctionManagerV2

[] Auction

[] Your BidderMetadata account

[signer optional/writable] Your Bidder account - Only needs to be signer if payer does not own

[signer] Payer

[] Token program

[] Token Vault program

[] Token metadata program

[] Store

[] System

[] Rent sysvar

[signer] Transfer authority to move the payment in the auction’s token_mint coin from the bidder account for the participation_fixed_price on the auction manager to the auction manager account itself.

[writable] The accept payment account for the auction manager

[writable] The token account you will potentially pay for the open edition bid with if necessary

[writable] Participation NFT printing holding account (present on participation_state)

[] Auction extended (pda relative to auction of [‘auction’, program id, vault key, ‘extended’])

StartAuction

If the auction manager is in Validated state, it can invoke the start command via calling this command here.

[writable] Auction manager

[writable] Auction

[signer] Auction manager authority

[] Store key

[] Auction program

[] Clock sysvar

ClaimBid

If the auction manager is in a Disbursing or Finished state, then this means Auction must be in Ended state. Then this end point can be used as a signed proxy to use auction manager’s authority over the auction to claim bid funds into the accept payment account on the auction manager for a given bid. Auction has no opinions on how bids are redeemed, only that they exist, have been paid, and have a winning place. It is up to the implementer of the auction to determine redemption, and auction manager does this via bid redemption tickets and the vault contract which ensure the user always can get their NFT once they have paid. Therefore, once they have paid, and the auction is over, the artist can claim funds at any time without any danger to the user of losing out on their NFT, because the AM will honor their bid with an NFT at ANY time.

[writable] The accept payment account on the auction manager

[writable] The bidder pot token account

[writable] The bidder pot pda account [seed of [‘auction’, program_id, auction key, bidder key] - relative to the auction program, not auction manager

[writable] Auction manager

[] The auction

[] The bidder wallet

[] Token mint of the auction

[] Vault

[] Store

[] Auction program

[] Clock sysvar

[] Token program

[] Auction extended (pda relative to auction of [‘auction’, program id, vault key, ‘extended’])

EmptyPaymentAccount(EmptyPaymentAccountArgs)

At any time, the auction manager authority may empty whatever funds are in the accept payment account on the auction manager. Funds come here from fixed price payments for partipation nfts, and from draining bid payments from the auction.

This action specifically takes a given safety deposit box, winning config, and creator on a metadata for the token inside that safety deposit box and pumps the requisite monies out to that creator as required by the royalties formula.

It’s up to the UI to iterate through all winning configs, all safety deposit boxes in a given winning config tier, and all creators for each metadata attached to each safety deposit box, to get all the money. Note that one safety deposit box can be used in multiple different winning configs, but this shouldn’t make any difference to this function.

We designed this function to be called in this loop-like manner because there is a limit to the number of accounts that can be passed up at once (32) and there may be many more than that easily in a given auction, so it’s easier for the implementer to just loop through and call it, and there is an incentive for them to do so (to get paid.) It’s permissionless as well as it will empty into any destination account owned by the creator that has the proper mint, so anybody can call it.

For the participation NFT, there is no winning config, but the total is figured by summing the winning bids and subtracting from the total escrow amount present.

[writable] The accept payment account on the auction manager

[writable] The destination account of same mint type as the accept payment account. Must be an Associated Token Account.

[writable] Auction manager

[writable] Payout ticket info to keep track of this artist or auctioneer’s payment, pda of [metaplex, auction manager, winning config index OR ‘participation’, safety deposit key]

[signer] payer

[] The metadata

[] The master edition of the metadata (optional if exists) (pda of [‘metadata’, program id, metadata mint id, ‘edition’]) - remember PDA is relative to token metadata program

[] Safety deposit box account

[] The store of the auction manager

[] The vault

[] Auction

[] Token program

[] System program

[] Rent sysvar

[] AuctionWinnerTokenTypeTracker, pda of seed [‘metaplex’, program id, auction manager key, ‘totals’]

[] Safety deposit config pda of [‘metaplex’, program id, auction manager, safety deposit]

Tuple Fields of EmptyPaymentAccount

0: EmptyPaymentAccountArgs

SetStore(SetStoreArgs)

Given a signer wallet, create a store with pda [‘metaplex’, wallet] (if it does not exist) and/or update it (if it already exists). Stores can be set to open (anybody can publish) or closed (publish only via whitelist).

[writable] The store key, seed of [‘metaplex’, admin wallet]

[signer] The admin wallet

[signer] Payer

[] Token program

[] Token vault program

[] Token metadata program

[] Auction program

[] System

[] Rent sysvar

Tuple Fields of SetStore

0: SetStoreArgs

SetWhitelistedCreator(SetWhitelistedCreatorArgs)

Given an existing store, add or update an existing whitelisted creator for the store. This creates a PDA with seed [‘metaplex’, store key, creator key] if it does not already exist to store attributes there.

[writable] The whitelisted creator pda key, seed of [‘metaplex’, store key, creator key]

[signer] The admin wallet

[signer] Payer

[] The creator key

[] The store key, seed of [‘metaplex’, admin wallet]

[] System

[] Rent sysvar

Tuple Fields of SetWhitelistedCreator

0: SetWhitelistedCreatorArgs

DeprecatedValidateParticipation

NOTE: Requires an AuctionManagerV1. Validates an participation nft (if present) on the Auction Manager. Because of the differing mechanics of an open edition (required for participation nft), it needs to be validated at a different endpoint than a normal safety deposit box. 0. [writable] Auction manager

[] Open edition metadata

[] Open edition MasterEdition account

[] Printing authorization token holding account - must be of the printing_mint type on the master_edition, used by the auction manager to hold printing authorization tokens for all eligible winners of the participation nft when auction ends. Must be owned by auction manager account.

[signer] Authority for the Auction Manager

[] A whitelisted creator entry for this store for the open edition pda of [‘metaplex’, store key, creator key] where creator key comes from creator list of metadata

[] The auction manager’s store

[] Safety deposit box

[] Safety deposit token store

[] Vault

[] Rent sysvar

DeprecatedPopulateParticipationPrintingAccount

NOTE: Requires an AuctionManagerV1. Needs to be called by someone at the end of the auction - will use the one time authorization token to fire up a bunch of printing tokens for use in participation redemptions.

[writable] Safety deposit token store

[writable] Transient account with mint of one time authorization account on master edition - you can delete after this txn

[writable] The printing token account on the participation state of the auction manager

[writable] One time printing authorization mint

[writable] Printing mint

[writable] Safety deposit of the participation prize

[writable] Vault info

[] Fraction mint

[] Auction info

[] Auction manager info

[] Token program

[] Token vault program

[] Token metadata program

[] Auction manager store

[] Master edition

[] PDA-based Transfer authority to move the tokens from the store to the destination seed [‘vault’, program_id] but please note that this is a PDA relative to the Token Vault program, with the ‘vault’ prefix

[] Payer who wishes to receive refund for closing of one time transient account once we’re done here

[] Rent

RedeemUnusedWinningConfigItemsAsAuctioneer(RedeemUnusedWinningConfigItemsAsAuctioneerArgs)

If you are an auctioneer, redeem an unused winning config entry. You provide the winning index, and if the winning index has no winner, then the correct redemption method is called with a special flag set to ignore bidder_metadata checks and a hardcoded winner index to empty this win to you.

All the keys, in exact sequence, should follow the expected call you wish to proxy to, because these will be passed to the process_ method of the next call. This method exists primarily to pass in an additional argument to the other redemption methods that subtly changes their behavior. We made this additional call so that if the auctioneer calls those methods directly, they still act the same as if the auctioneer were a normal bidder, which is be desirable behavior.

An auctioneer should never be in the position where the auction can never work the same for them simply because they are an auctioneer. This special endpoint exists to give them the “out” to unload items via a proxy call once the auction is over.

Tuple Fields of RedeemUnusedWinningConfigItemsAsAuctioneer

0: RedeemUnusedWinningConfigItemsAsAuctioneerArgs

DecommissionAuctionManager

If you have an auction manager in an Initialized state and for some reason you can’t validate it, you want to retrieve The items inside of it. This will allow you to move it straight to Disbursing, and then you can, as Auctioneer, Redeem those items using the RedeemUnusedWinningConfigItemsAsAuctioneer endpoint.

If you pass the vault program account, authority over the vault will be returned to you, so you can unwind the vault to get your items back that way instead.

Be WARNED: Because the boxes have not been validated, the logic for redemptions may not work quite right. For instance, if your validation step failed because you provided an empty box but said there was a token in it, when you go and try to redeem it, you yourself will experience quite the explosion. It will be up to you to tactfully request the bids that can be properly redeemed from the ones that cannot.

If you had a FullRightsTransfer token, and you never validated (and thus transferred) ownership, when the redemption happens it will skip trying to transfer it to you, so that should work fine.

[writable] Auction Manager

[writable] Auction

[signer] Authority of the Auction Manager

[writable] Vault

[] Store

[] Auction program

[] Clock sysvar

[] Vault program (Optional)

RedeemPrintingV2Bid(RedeemPrintingV2BidArgs)

Note: This requires that auction manager be in a Running state and that be of the V1 type.

If an auction is complete, you can redeem your printing v2 bid for a specific item here. If you are the first to do this, The auction manager will switch from Running state to Disbursing state. If you are the last, this may change the auction manager state to Finished provided that no authorities remain to be delegated for Master Edition tokens.

NOTE: Please note that it is totally possible to redeem a bid 2x - once for a prize you won and once at the RedeemParticipationBid point for an open edition that comes as a ‘token of appreciation’ for bidding. They are not mutually exclusive unless explicitly set to be that way.

[writable] Auction manager

[writable] Safety deposit token storage account

[writable] Account containing 1 token of your new mint type. MUST be an associated token account of pda [wallet, token program, mint] relative to ata program.

[writable] Bid redemption key - Just a PDA with seed [‘metaplex’, auction_key, bidder_metadata_key] that we will allocate to mark that you redeemed your bid

[writable] Safety deposit box account

[writable] Vault account

[] Safety deposit config pda of [‘metaplex’, program id, auction manager, safety deposit] This account will only get used in the event this is an AuctionManagerV2

[] Auction

[] Your BidderMetadata account

[] Your Bidder account - Only needs to be signer if payer does not own

[signer] Payer

[] Token program

[] Token Vault program

[] Token metadata program

[] Store

[] System

[] Rent sysvar

[writable] Prize tracking ticket (pda of [‘metaplex’, program id, auction manager key, metadata mint id])

[writable] New Metadata key (pda of [‘metadata’, program id, mint id])

[writable] New Edition (pda of [‘metadata’, program id, mint id, ‘edition’])

[writable] Master Edition of token in vault V2 (pda of [‘metadata’, program id, master metadata mint id, ‘edition’]) PDA is relative to token metadata.

[writable] Mint of new token

[writable] Edition pda to mark creation - will be checked for pre-existence. (pda of [‘metadata’, program id, master metadata mint id, ‘edition’, edition_number]) where edition_number is NOT the edition number you pass in args but actually edition_number = floor(edition/EDITION_MARKER_BIT_SIZE). PDA is relative to token metadata.

[signer] Mint authority of new mint - THIS WILL TRANSFER AUTHORITY AWAY FROM THIS KEY

[] Metadata account of token in vault

[] Auction extended (pda relative to auction of [‘auction’, program id, vault key, ‘extended’])

Tuple Fields of RedeemPrintingV2Bid

0: RedeemPrintingV2BidArgs

WithdrawMasterEdition

Permissionless call to redeem the master edition in a given safety deposit for a PrintingV2 winning config to the ATA of the Auctioneer. Can only be called once all redemptions have been met.

[writable] Auction manager

[writable] Safety deposit token storage account

[writable] Associated token account owned by auction manager authority of same mint as token storage account

[writable] Safety deposit box account

[writable] Vault account

[writable] Fraction mint of the vault

[] Prize tracking ticket (pda of [‘metaplex’, program id, auction manager key, metadata mint id])

[] PDA-based Vault transfer authority [‘vault’, program_id, vault key] but please note that this is a PDA relative to the Token Vault program, with the ‘vault’ prefix

[] Auction

[] Auction data extended (pda relative to auction of [‘auction’, program id, vault key, ‘extended’])

[] Token program

[] Token Vault program

[] Store

[] Rent sysvar

[] Safety deposit config pda of [‘metaplex’, program id, auction manager, safety deposit] This account will only get used in the event this is an AuctionManagerV2

DeprecatedRedeemParticipationBidV2

Note: This requires that auction manager be in a Running state.

Second note: Unlike it’s predecessor, V2 is permissionless. You can in theory pay for someone else’s participation NFT and gift it to them.

If an auction is complete, you can redeem your bid for an Open Edition token if it is eligible. If you are the first to do this, The auction manager will switch from Running state to Disbursing state. If you are the last, this may change the auction manager state to Finished provided that no authorities remain to be delegated for Master Edition tokens.

NOTE: Please note that it is totally possible to redeem a bid 2x - once for a prize you won and once at this end point for a open edition that comes as a ‘token of appreciation’ for bidding. They are not mutually exclusive unless explicitly set to be that way.

NOTE: If you are redeeming a newly minted Open Edition, you must actually supply a destination account containing a token from a brand new mint. We do not provide the token to you. Our job with this action is to christen this mint + token combo as an official Open Edition.

[writable] Auction manager

[writable] Safety deposit token storage account

[writable] Account containing 1 token of your new mint type. MUST be an associated token account of pda [wallet, token program, mint] relative to ata program.

[writable] Bid redemption key - Just a PDA with seed [‘metaplex’, auction_key, bidder_metadata_key] that we will allocate to mark that you redeemed your bid

[] Safety deposit box account

[] Vault account

[writable] Safety deposit config pda of [‘metaplex’, program id, auction manager, safety deposit] This account will only get used in the event this is an AuctionManagerV2

[] Auction

[] Your BidderMetadata account

[] Your Bidder account

[signer] Payer

[] Token program

[] Token Vault program

[] Token metadata program

[] Store

[] System

[] Rent sysvar

[signer] Transfer authority to move the payment in the auction’s token_mint coin from the bidder account for the participation_fixed_price on the auction manager to the auction manager account itself.

[writable] The accept payment account for the auction manager

[writable] The token account you will potentially pay for the open edition bid with if necessary.

[writable] Prize tracking ticket (pda of [‘metaplex’, program id, auction manager key, metadata mint id])

[writable] New Metadata key (pda of [‘metadata’, program id, mint id])

[writable] New Edition (pda of [‘metadata’, program id, mint id, ‘edition’])

[writable] Master Edition of token in vault V2 (pda of [‘metadata’, program id, master metadata mint id, ‘edition’]) PDA is relative to token metadata.

[writable] Mint of new token

[writable] Edition pda to mark creation - will be checked for pre-existence. (pda of [‘metadata’, program id, master metadata mint id, ‘edition’, edition_number]) where edition_number is NOT the edition number you pass in args but actually edition_number = floor(edition/EDITION_MARKER_BIT_SIZE). PDA is relative to token metadata.

[signer] Mint authority of new mint - THIS WILL TRANSFER AUTHORITY AWAY FROM THIS KEY

[] Metadata account of token in vault

InitAuctionManagerV2(InitAuctionManagerV2Args)

Initializes an Auction Manager V2

NOTE: It is not possible to use MasterEditionV1s for participation nfts with these managers.

[writable] Uninitialized, unallocated auction manager account with pda of [‘metaplex’, auction_key from auction referenced below]

[writable] AuctionWinnerTokenTypeTracker, pda of seed [‘metaplex’, program id, auction manager key, ‘totals’]

[] Combined vault account with authority set to auction manager account (this will be checked) Note in addition that this vault account should have authority set to this program’s pda of [‘metaplex’, auction_key]

[] Auction with auctioned item being set to the vault given and authority set to this program’s pda of [‘metaplex’, auction_key]

[] Authority for the Auction Manager

[signer] Payer

[] Accept payment account of same token mint as the auction for taking payment for open editions, owner should be auction manager key

[] Store that this auction manager will belong to

[] System sysvar

[] Rent sysvar

Tuple Fields of InitAuctionManagerV2

0: InitAuctionManagerV2Args

ValidateSafetyDepositBoxV2(SafetyDepositConfig)

NOTE: Requires an AuctionManagerV2.

Validates that a given safety deposit box has in it contents that match the given SafetyDepositConfig, and creates said config. A stateful call, this will error out if you call it a second time after validation has occurred. 0. [writable] Uninitialized Safety deposit config, pda of seed [‘metaplex’, program id, auction manager key, safety deposit key]

[writable] AuctionWinnerTokenTypeTracker, pda of seed [‘metaplex’, program id, auction manager key, ‘totals’]

[writable] Auction manager

[writable] Metadata account

[writable] Original authority lookup - unallocated uninitialized pda account with seed [‘metaplex’, auction key, metadata key] We will store original authority here to return it later.

[] A whitelisted creator entry for the store of this auction manager pda of [‘metaplex’, store key, creator key] where creator key comes from creator list of metadata, any will do

[] The auction manager’s store key

[] Safety deposit box account

[] Safety deposit box storage account where the actual nft token is stored

[] Mint account of the token in the safety deposit box

[] Edition OR MasterEdition record key Remember this does not need to be an existing account (may not be depending on token), just is a pda with seed of [‘metadata’, program id, Printing mint id, ‘edition’]. - remember PDA is relative to token metadata program.

[] Vault account

[signer] Authority

[signer optional] Metadata Authority - Signer only required if doing a full ownership txfer

[signer] Payer

[] Token metadata program

[] System

[] Rent sysvar

Tuple Fields of ValidateSafetyDepositBoxV2

0: SafetyDepositConfig

RedeemParticipationBidV3(RedeemParticipationBidV3Args)

Note: This requires that auction manager be in a Running state.

Second note: V3 is the same as V2, but it requires an additional argument because it is intended to be used with AuctionManagerV2s, not V1s, which use BidRedemptionTicketV2s, which require this additional argument (the user_provided_win_index). You can in theory pay for someone else’s participation NFT and gift it to them.

If an auction is complete, you can redeem your bid for an Open Edition token if it is eligible. If you are the first to do this, The auction manager will switch from Running state to Disbursing state. If you are the last, this may change the auction manager state to Finished provided that no authorities remain to be delegated for Master Edition tokens.

NOTE: Please note that it is totally possible to redeem a bid 2x - once for a prize you won and once at this end point for a open edition that comes as a ‘token of appreciation’ for bidding. They are not mutually exclusive unless explicitly set to be that way.

NOTE: If you are redeeming a newly minted Open Edition, you must actually supply a destination account containing a token from a brand new mint. We do not provide the token to you. Our job with this action is to christen this mint + token combo as an official Open Edition.

[writable] Auction manager

[writable] Safety deposit token storage account

[writable] Account containing 1 token of your new mint type. MUST be an associated token account of pda [wallet, token program, mint] relative to ata program.

[writable] Bid redemption key - Just a PDA with seed [‘metaplex’, auction_key, bidder_metadata_key] that we will allocate to mark that you redeemed your bid

[] Safety deposit box account

[] Vault account

[writable] Safety deposit config pda of [‘metaplex’, program id, auction manager, safety deposit] This account will only get used in the event this is an AuctionManagerV2

[] Auction

[] Your BidderMetadata account

[] Your Bidder account

[signer] Payer

[] Token program

[] Token Vault program

[] Token metadata program

[] Store

[] System

[] Rent sysvar

[signer] Transfer authority to move the payment in the auction’s token_mint coin from the bidder account for the participation_fixed_price on the auction manager to the auction manager account itself.

[writable] The accept payment account for the auction manager

[writable] The token account you will potentially pay for the open edition bid with if necessary.

[writable] Prize tracking ticket (pda of [‘metaplex’, program id, auction manager key, metadata mint id])

[writable] New Metadata key (pda of [‘metadata’, program id, mint id])

[writable] New Edition (pda of [‘metadata’, program id, mint id, ‘edition’])

[writable] Master Edition of token in vault V2 (pda of [‘metadata’, program id, master metadata mint id, ‘edition’]) PDA is relative to token metadata.

[writable] Mint of new token

[writable] Edition pda to mark creation - will be checked for pre-existence. (pda of [‘metadata’, program id, master metadata mint id, ‘edition’, edition_number]) where edition_number is NOT the edition number you pass in args but actually edition_number = floor(edition/EDITION_MARKER_BIT_SIZE). PDA is relative to token metadata.

[signer] Mint authority of new mint - THIS WILL TRANSFER AUTHORITY AWAY FROM THIS KEY

[] Metadata account of token in vault

Tuple Fields of RedeemParticipationBidV3

0: RedeemParticipationBidV3Args

EndAuction(EndAuctionArgs)

Ends an auction, regardless of end timing conditions.

[writable] Auction manager

[writable] Auction

[] Auction extended data account (pda relative to auction of [‘auction’, program id, vault key, ‘extended’]).

[signer] Auction manager authority

[] Store key

[] Auction program

[] Clock sysvar

Tuple Fields of EndAuction

0: EndAuctionArgs

SetStoreIndex(SetStoreIndexArgs)

Creates/Updates a store index page

[writable] Store index (pda of [‘metaplex’, program id, store key, ‘index’, page_number])

[signer] Payer info

[] Auction cache (pda of [‘metaplex’, program id, store key, auction key, ‘cache’])

[] Store key

[] System

[] Rent sysvar

[optional] Auction cache above current (pda of [‘metaplex’, program id, store key, auction key, ‘cache’]) Note: Can pass the below in this slot if there is no above

[optional] Auction cache below current (pda of [‘metaplex’, program id, store key, auction key, ‘cache’])

Tuple Fields of SetStoreIndex

0: SetStoreIndexArgs

SetAuctionCache

Creates/Updates a store index page

[writable] Auction cache (pda of [‘metaplex’, program id, store key, auction key, ‘cache’])

[signer] Payer info

[] Auction

[] Safety deposit box account

[] Auction manager

[] Store key

[] System

[] Rent sysvar

[] Clock sysvar

SetStoreV2(SetStoreV2Args)

Given a signer wallet, create a store with pda [‘metaplex’, wallet] (if it does not exist) and/or update it (if it already exists). Stores can be set to open (anybody can publish) or closed (publish only via whitelist).

[writable] The store key, seed of [‘metaplex’, admin wallet]

[writable] The store config key, seed of [‘metaplex’, store key]

[signer] The admin wallet

[signer] Payer

[] Token program

[] Token vault program

[] Token metadata program

[] Auction program

[] System

[] Rent sysvar

Tuple Fields of SetStoreV2

0: SetStoreV2Args